backend/requirements.txt
fastapi==0.117.1
pydantic==2.11.9
python-multipart==0.0.20
uvicorn==0.37.0
pandas==2.3.2
numpy>=1.23.2
gunicorn==23.0.0
google-cloud-bigquery>=3.26.0
google-cloud-bigquery-storage>=2.26.0
cachetools>=5.3.0
pyarrow>=16.1.0
pytest>=8.3.0
httpx>=0.27.0